Hacker Concept

+++ WeaponsControl/Scripts/Ammo.cs
@@ -60,7 +60,7 @@ namespace Rynchodon.Weapons
 			/// <summary>Detonate when this close to target.</summary>
 			public float DetonateRange = 0f;
 
-			public float EMP_Strength = 0f;
+			public int EMP_Strength = 0;
 			public float EMP_Seconds = 0f;
 
 			#region Cluster
